Add a generic cli parsing function
This is a bit of a distraction.
It's there because I wanted to generically parse cli arguments for all
the builders using a map of json field names to builder function names.
It was a ridiculous idea for a project on this scale. Included for
completeness. Perhaps it will be useful in a more generic "setup
builders from clap cli arg values" thing, but perhaps not.